+++ synergy/higher/synergy_model_Nd.py
@@ -14,7 +14,7 @@ from synergy.single.dose_response_model_1d import DoseResponseModel1D
 from synergy.utils import dose_utils
 from synergy.utils.model_mixins import ParametricModelMixins
 
-_LOGGER = logging.Logger(__name__)
+_LOGGER = logging.getLogger(__name__)
 
 
 class SynergyModelND(ABC):